requirements.txt
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-modeladmin-reorder==0.3.1
djangorestframework==3.15.2
inflection==0.5.1
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
requests==2.32.3
setuptools==70.1.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2